application uses gradle build system
